Create CoreConst from HygroConst
